renamed speechToText.ts
