+++ frontend/src/components/query-input/speechToText.ts
